NeuroSpike-MNIST/requirements.txt
torch>=1.9.0
torchvision>=0.10.0
snntorch>=0.6.0
numpy>=1.24.3
matplotlib>=3.4.0
seaborn>=0.11.0
pyyaml>=5.4.1
tqdm>=4.62.0
pytest>=6.2.5
flake8>=3.9.0
tensorboard>=2.7.0